+++ web/src/rendered_markdown.ts
@@ -345,7 +345,21 @@ export const update_elements = ($content: JQuery): void => {
             const $inline_img_thumbnail = $(this);
             let thumbnail_name = thumbnail.preferred_format.name;
             if ($inline_img_thumbnail.attr("data-animated") === "true") {
-                thumbnail_name = thumbnail.animated_format.name;
+                if (
+                    user_settings.web_animate_image_previews === "always" ||
+                    // Treat on_hover as "always" on mobile web, where
+                    // hovering is impossible and there's much less on
+                    // the screen.
+                    (user_settings.web_animate_image_previews === "on_hover" && util.is_mobile())
+                ) {
+                    thumbnail_name = thumbnail.animated_format.name;
+                } else {
+                    // If we're showing a still thumbnail, show a play
+                    // button so that users that it can be played.
+                    $inline_img_thumbnail
+                        .closest(".message_inline_image")
+                        .addClass("message_inline_animated_image_still");
+                }
             }
             $inline_img_thumbnail.attr(
                 "src",