+++ zerver/models/users.py
@@ -89,6 +89,9 @@ class UserBaseSettings(models.Model):
         default=WEB_LINE_HEIGHT_PERCENT_DEFAULT
     )
 
+    # UI setting to control how animated images are played.
+    web_animate_image_previews = models.TextField(default="on_hover")
+
     # UI setting controlling Zulip's behavior of demoting in the sort
     # order and graying out streams with no recent traffic.  The
     # default behavior, automatic, enables this behavior once a user
@@ -350,6 +353,7 @@ class UserBaseSettings(models.Model):
         web_mark_read_on_scroll_policy=int,
         web_channel_default_view=int,
         user_list_style=int,
+        web_animate_image_previews=str,
         web_stream_unreads_count_display_policy=int,
         web_font_size_px=int,
         web_line_height_percent=int,